"Who we are" panel
Added the section on our team, in its first iteration. The section about the
team members will be changed to a rail in the future.

Additionally some minor tweaks to other parts of the page, e.g. adding the :focus
state to all links, so that they are tabable again.